Extracted ClaimPathBuilder and FactionInfoBuilder from MfDynmapService
